Added post search handler and search view
The searchHandler checks some cases such as post not found, id less
than 0 and id not being a string if the form input does not catch
it somehow. Then renders the post view instead of the old view
handler. If there are any errors, the search template has some ifs
to add error messages to the search view. I need to still remove or
handle the old post view.